GH-642 regroupped tests
